pyproject.toml
[tool.poetry]
name = "next-step-agent"
version = "0.1.0"
description = ""
authors = ["Aubrey <aubrey@rainsound.ai>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.0"
python-dateutil = "^2.8.2"
google-auth-httplib2 = "^0.2.0"
google-api-python-client = "^2.154.0"
google-auth-oauthlib = "^1.2.1"
slack-sdk = "^3.21.3"
openai = "^1.3.0"
httpx = "^0.25.0"
pytz = "^2023.3.post1"
numpy = "^1.24.3"
websockets = "^11.0.3"
notion-client = "^2.2.1"
flask = "^3.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
